code/requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==9.5.0